serve_cars -> float

+++ app/main.py
@@ -20,7 +20,7 @@ class CarWashStation:
         self.average_rating = average_rating
         self.count_of_ratings = count_of_ratings
 
-    def serve_cars(self, cars: list[Car]) -> int:
+    def serve_cars(self, cars: list[Car]) -> float:
         price = 0.0
         for auto in cars:
             if auto.clean_mark < self.clean_power:
